+++ src/app/main/pages/main/main.component.ts
@@ -1,17 +1,15 @@
-import { ChangeDetectorRef,
-         Component,
-         OnInit,
-         NgZone } from '@angular/core';
-import { AuthService } from 'src/app/auth/services/auth.service';
-
+import { ChangeDetectorRef, Component, OnInit, NgZone, Inject} from '@angular/core';
 import { Idle, DEFAULT_INTERRUPTSOURCES } from '@ng-idle/core';
+import { AuthService } from 'src/app/auth/services/auth.service';
 import { MenuItem } from 'primeng/api';
 import { SocketService } from 'src/app/services/socket.service';
 import { Logout } from 'src/app/auth/interfaces/Logout';
 import { GlobalService } from 'src/app/services/global.service';
 import { Router } from '@angular/router';
-import { MainService } from '../../services/main.service';
+import { DOCUMENT } from '@angular/common';
+
 import * as moment from 'moment';
+import { ChabotService } from '../../chat-bot/services/chatbot.service';
 moment.locale("es");
 
 @Component({
@@ -21,40 +19,37 @@ moment.locale("es");
 })
 export class MainComponent implements OnInit {
 
+  /** variables de clase */
   themeSelecion = false;
   display:boolean = false;
   ruta: MenuItem[];
 
-  //representa la cuenta regresiva del modal
+  /** representa la cuenta regresiva del modal */
   contador:number = 20;
-  //representa el tiempo en segundos de intervalo permitido de inactividad en el sistema
-  //240s representa 10 min de holgura
+  /** representa el tiempo en segundos de intervalo permitido de inactividad en el sistema
+  * 240s representa 10 min de holgura
+  */
   tiempo:number = 600;
   countdown?: number;
   lastPing?: Date;
 
-  constructor(private route:Router,
-              private _auth:AuthService,
-              public _global:GlobalService,
-              private idle: Idle,
-              private ngZone: NgZone,
-              private cd: ChangeDetectorRef,
-              private _socket:SocketService) {
-                this.InitIdle();
-                this.welcomeMessage();
-              }
+  constructor(@Inject(DOCUMENT)
+              private readonly document:Document,
+              private readonly route:Router,
+              private readonly _auth:AuthService,
+              public  readonly _global:GlobalService,
+              private readonly idle: Idle,
+              private readonly ngZone: NgZone,
+              private readonly cd: ChangeDetectorRef,
+              private readonly _socket:SocketService,
+              private readonly _chatbot:ChabotService){
 
-  welcomeMessage(){
-    this._socket.OnEvent('welcome-ceidbot').subscribe({
-      next: (value) => {
-        console.log(value);
-      },
-      error: (err) => console.log(err)
-    })
+                this.InitIdle();
+                this._socket.Onconexion();
   }
 
   reset() {
-    //TODO: por defecto inicia en NOT_IDLE
+    //? TODO: por defecto inicia en NOT_IDLE
     this.idle.watch();
     this.countdown = undefined;
     this.lastPing = undefined;
@@ -64,6 +59,12 @@ export class MainComponent implements OnInit {
     this.reset();
   }
 
+  modeTheme(value:boolean, system:boolean = true){
+    const themeLink = this.document.getElementById('app-theme') as HTMLLinkElement;
+    themeLink.href = value?'mdc-dark-indigo.css':'mdc-light-indigo.css'
+    if(system) this._global.saveTheme(value);
+  }
+
   InitIdle(){
     // set idle parameters
     this.idle.setIdle(this.tiempo);
@@ -71,18 +72,18 @@ export class MainComponent implements OnInit {
     this.idle.setInterrupts(DEFAULT_INTERRUPTSOURCES);
 
     this.idle.onIdleStart.subscribe(() => {
-      //TODO: se activa el contador setTimeout()
+      //? TODO: se activa el contador setTimeout()
       console.log("cerrando sesión en ", this.countdown);
     });
 
     this.idle.onIdleEnd.subscribe(() => {
-      //TODO: se activa cuando el usurio reinicia la tarea en el ordenador
+      //? TODO: se activa cuando el usurio reinicia la tarea en el ordenador
       this.countdown = undefined;
       this.cd.detectChanges();
     });
 
     this.idle.onTimeout.subscribe(() =>{
-      //TODO: termina el observer
+      //? TODO: termina el observer
       console.log("Cerrando sesión por seguridad");
       this.ngZone.run(() => {
         this.logout();
@@ -101,16 +102,12 @@ export class MainComponent implements OnInit {
     const data = new Logout(this._auth.userAuth?.Id, this._auth.userAuth?.Email);
     this._auth.logout(data).subscribe({
       next:(value) => {
-        if(value.ok){
-          if(this._auth.readToken()){
-            this._auth.deleteToken();
-          }
-          this.route.navigate(['/main/auth/login']);
-        }
+        if(!value.ok) return;
+        if(this._auth.readToken()) this._auth.deleteToken();
+        this.modeTheme(false, false);
+        this.route.navigate(['/main/auth/login']);
       },
-      error:(err) => {
-        console.log("cerrar sesión",err);
-      },
+      error:(err) => console.log("cerrar sesión",err)
     })
 
   }
